requirements.txt
# Core Django
Django==6.0.1
psycopg2-binary==2.9.9
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3

# Database & GIS
django-contrib-postgres==0.1.0
django-leaflet==0.30.1
django-map-widgets==0.4.2
GDAL==3.4.3

# Storage & Media
django-storages==1.14.2
boto3==1.34.7
Pillow==10.1.0

# Forms & UI
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-tinymce==3.6.0
django-ckeditor==6.7.0

# Authentication
django-allauth==0.57.0
django-otp==1.2.0  # For 2FA
qrcode==7.4.2

# REST API
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
drf-yasg==1.21.7  # API documentation
markdown==3.5.1

# Background Tasks
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
flower==2.0.1  # Celery monitoring

# Search
django-haystack==3.2.1
whoosh==2.7.4
elasticsearch==7.13.4

# Admin & Debug
django-admin-interface==0.22.4
django-colorfield==0.9.0
django-debug-toolbar==4.2.0
django-extensions==3.2.3
ipython==8.18.1

# Import/Export
django-import-export==3.3.1
openpyxl==3.1.2
tablib==3.5.0

# SEO & Sitemaps
django-robots==5.0
django-sitemaps==0.7
django-taggit==4.0.0
django-hitcount==1.3.5

# Payments (M-Pesa)
requests==2.31.0
django-mpesa==0.2.0
django-payments==1.0.0

# Analytics
sentry-sdk==1.39.0
django-google-analytics-app==5.1
mixpanel==4.10.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==20.1.0
model-bakery==1.17.0

# Performance
django-cacheops==7.0.1
django-compressor==4.4
django-htmlmin==0.11.0
django-optimizations==1.0.0

# Internationalization
django-modeltranslation==0.18.11
django-countries==7.5.1

# Async & WebSockets
channels==4.0.0
channels-redis==4.1.0
websockets==12.0

# Utilities
python-dateutil==2.8.2
django-cleanup==8.0.0
django-js-asset==2.1.0
django-widget-tweaks==1.5.0
pillow==10.1.0